Find surface now ignores the walls

+++ heat_transfer_equation.py
@@ -60,8 +60,8 @@ Returns:
 '''
 @jit
 def hte_calculate(n_x, n_y, n_z, surface, delta_T_0, temperature, Lambda, Dr, dx, dy, dz, dt, density, heat_capacity, j_leave, j_inward, latent_heat_water, j_leave_co2, j_inward_co2, latent_heat_co2):
-    delta_T = np.zeros((n_x, n_y, n_z)) + delta_T_0
-    Energy_Increase_per_Layer = np.zeros(n_x, n_y, n_z)
+    delta_T = np.zeros((n_x, n_y, n_z), dtype=np.float64) + delta_T_0
+    Energy_Increase_per_Layer = np.zeros((n_x, n_y, n_z), dtype=np.float64)
     #Latent_Heat_per_Layer = np.zeros(n_x, n_y, n_z)
     #Fourier_number = np.zeros(n_x, n_y, n_z)
     for i in range(1, n_z):
